moved SentenceType tp Nomis module where it belongs
